pyproject.toml
[tool.poetry]
name = "docsend"
version = "0.2.0"
description = "convert docsend to pdf or png sequence"
authors = ["banteg <banteeg@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.0"
requests-html = "^0.9.0"
pillow = "^5.3"
pymupdf = "^1.23.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
docsend = "docsend.cli:main"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

